==> docs/3b. OLS Model.Rmd <==
---
title: "OLS Model 2016"
output: html_notebook
---

## Introduction
This notebook is one of a multiple part series for analyzing csb and voter data. This notebook accomplishes the building of the OLS model using the python library `pysal` and the R package `reticulate` to be able to execute python code in an R notebook.

This notebook fits models using 2016 election data.

## Dependencies

### Software Versions
R Version 3.5.2
RStudio Version 1.2.1330
Python Version 3.7.3
Pysal Version 2.0.0
Numpy Version 1.15.3

These are the R packages we need:
```{r dependencies}
library(reticulate) # python interface
```

And these are the Python libraries we need:

```{python imports}
import os
import pysal as ps
import numpy as np
```

## Building the OLS Models
### Create Spatial Weights
We'll use queens weights.
```{python}
file = ps.lib.io.open("../data/Spatial/grid/full_16.shp")
w = ps.lib.weights.Queen(file)

```

### Set Variables for Models
We're going to create 4 models, all with the number of CSB calls as the dependent variable.

1. Main Effect (Voter Turnout)
2. Population Model (Population) 
3. Combined Model (Voter Turnout and Population)
4. Full Model (Voter Turnout, Population, Non-White Population, Poverty, Highschool Diploma, Bachelors Degree)

```{python}
# load data
data = ps.lib.io.open("../data/Spatial/grid/full_16.dbf", 'r')

# set Y var
y_name = "csb"

# create Y array, transpose
y = np.array([data.by_col(y_name)]).T 

# define independent vars for each model
me_ind = ["turnout"]
pm_ind = ["tot_pop"]
cm_ind = ["turnout", "tot_pop"]
fm_ind = ["turnout", "tot_pop", "nonwht", "pvrty", "hs", "ba"]

# create arrays for each model
me = np.array([data.by_col(var) for var in me_ind]).T
pm = np.array([data.by_col(var) for var in pm_ind]).T
cm = np.array([data.by_col(var) for var in cm_ind]).T
fm = np.array([data.by_col(var) for var in fm_ind]).T

```

## Create OLS Models
### Main Effect
```{python}
ols_me = ps.model.spreg.OLS(y, me, w=w, name_y=y_name, name_x=me_ind, spat_diag=True, moran=True,
name_w='queens', name_ds='full_16.shp')

print(ols_me.summary)
```

### Population Model
```{python}
ols_pm = ps.model.spreg.OLS(y, pm, w=w, name_y=y_name, name_x=pm_ind, spat_diag=True, moran=True,
name_w='queens', name_ds='full_16.shp')

print(ols_pm.summary)
```

### Combined Model
```{python}
ols_cm = ps.model.spreg.OLS(y, cm, w=w, name_y=y_name, name_x=cm_ind, spat_diag=True, moran=True,
name_w='queens', name_ds='full_16.shp')

print(ols_cm.summary)
```

### Full Model
```{python}
ols_fm = ps.model.spreg.OLS(y, fm, w=w, name_y=y_name, name_x=fm_ind, spat_diag=True, moran=True,
name_w='queens', name_ds='full_16.shp')

print(ols_fm.summary)
```
